add colour theme and remove duplicate files
